Rename integration test classes to include 'IntegTest' in class name
