+++ photon-api/src/integTest/scala/com/linkedin/photon/ml/optimization/RandomEffectOptimizationTrackerIntegTest.scala
@@ -26,7 +26,7 @@ import com.linkedin.photon.ml.util.FunctionValuesConverged
 /**
  * Some simple tests for random effect optimization tracker.
  */
-class RandomEffectOptimizationTrackerTest extends SparkTestUtils {
+class RandomEffectOptimizationTrackerIntegTest extends SparkTestUtils {
 
   @Test
   def testCountConvergenceReasons(): Unit = sparkTest("testCountConvergenceReasons") {
@@ -44,7 +44,7 @@ class RandomEffectOptimizationTrackerTest extends SparkTestUtils {
     val emptyOptimizationStatesTracker = new OptimizationStatesTracker()
     val normalOptimizationStatesTracker = new OptimizationStatesTracker()
     val numStates = 5
-    RandomEffectOptimizationTrackerTest.getDummyOptimizerStates(numStates)
+    RandomEffectOptimizationTrackerIntegTest.getDummyOptimizerStates(numStates)
         .foreach(normalOptimizationStatesTracker.track)
     val optimizationStatesTrackerAsRDD =
       sc.parallelize(Seq(emptyOptimizationStatesTracker, normalOptimizationStatesTracker))
@@ -67,7 +67,7 @@ class RandomEffectOptimizationTrackerTest extends SparkTestUtils {
   }
 }
 
-object RandomEffectOptimizationTrackerTest {
+object RandomEffectOptimizationTrackerIntegTest {
 
   /**
    *
