+++ photon-client/src/integTest/scala/com/linkedin/photon/ml/data/avro/ScoreProcessingUtilsIntegTest.scala
@@ -24,7 +24,7 @@ import com.linkedin.photon.ml.test.{SparkTestUtils, TestTemplateWithTmpDir}
 /**
  *
  */
-class ScoreProcessingUtilsTest extends SparkTestUtils with TestTemplateWithTmpDir {
+class ScoreProcessingUtilsIntegTest extends SparkTestUtils with TestTemplateWithTmpDir {
 
   private val completeScoreItems = Array(
     ScoredItem(
