+++ photon-lib/src/integTest/scala/com/linkedin/photon/ml/data/scoring/CoordinateDataScoresIntegTest.scala
@@ -20,57 +20,57 @@ import org.testng.annotations.Test
 import com.linkedin.photon.ml.test.SparkTestUtils
 
 /**
- * Simple tests for [[ModelDataScores]].
+ * Simple tests for [[CoordinateDataScores]].
  */
-class ModelDataScoresTest extends SparkTestUtils {
+class CoordinateDataScoresIntegTest extends SparkTestUtils {
 
   /**
-   * Generate a [[ModelDataScores]] instance with the given scores.
+   * Generate a [[CoordinateDataScores]] instance with the given scores.
    *
    * @param scores The scores
-   * @return The [[ModelDataScores]] generated with the given scores
+   * @return The [[CoordinateDataScores]] generated with the given scores
    */
-  private def generateFullDataScores(scores: Array[Double]): ModelDataScores =
-    new ModelDataScores(sc.parallelize(scores.zipWithIndex.map { case (score, uniqueId) =>
-      (uniqueId.toLong, ScoredGameDatum(score = score))
+  private def generateSimpleDataScores(scores: Array[Double]): CoordinateDataScores =
+    new CoordinateDataScores(sc.parallelize(scores.zipWithIndex.map { case (score, uniqueId) =>
+      (uniqueId.toLong, score)
     }))
 
   /**
-   * Generate a [[ModelDataScores]] instance with the given keys and values.
+   * Generate a [[CoordinateDataScores]] instance with the given keys and values.
    *
    * @param keys The given keys
    * @param values The given values
-   * @return The [[ModelDataScores]] generated with the given keys and values
+   * @return The [[CoordinateDataScores]] generated with the given keys and values
    */
-  private def generateFullDataScores(keys: Array[Long], values: Array[Double]): ModelDataScores =
-    new ModelDataScores(sc.parallelize(keys.zip(values.map(v => ScoredGameDatum(score = v)))))
+  private def generateSimpleDataScores(keys: Array[Long], values: Array[Double]): CoordinateDataScores =
+    new CoordinateDataScores(sc.parallelize(keys.zip(values)))
 
   @Test
   def testEquals(): Unit = sparkTest("testEqualsForKeyValueScores") {
     //case 1: key value scores of length 0
-    val emptyScores1 = generateFullDataScores(scores = Array[Double]())
-    val emptyScores2 = generateFullDataScores(scores = Array[Double]())
+    val emptyScores1 = generateSimpleDataScores(scores = Array[Double]())
+    val emptyScores2 = generateSimpleDataScores(scores = Array[Double]())
     assertEquals(emptyScores1, emptyScores2)
 
     //case 2: key value scores with different length
-    val nonEmptyScores1 = generateFullDataScores(scores = Array[Double](1, 2))
-    val nonEmptyScores2 = generateFullDataScores(scores = Array[Double](1, 2, 3))
+    val nonEmptyScores1 = generateSimpleDataScores(scores = Array[Double](1, 2))
+    val nonEmptyScores2 = generateSimpleDataScores(scores = Array[Double](1, 2, 3))
     assertNotEquals(nonEmptyScores1, nonEmptyScores2)
 
     //case 3: key value scores with same keys but different values
     val sharedKeys = Array[Long](1, 3, 5)
     val values1 = Array[Double](0, 0, 0)
     val values2 = Array[Double](0, 1, 0)
-    val keyValueScore1 = generateFullDataScores(sharedKeys, values1)
-    val keyValueScore2 = generateFullDataScores(sharedKeys, values2)
+    val keyValueScore1 = generateSimpleDataScores(sharedKeys, values1)
+    val keyValueScore2 = generateSimpleDataScores(sharedKeys, values2)
     assertNotEquals(keyValueScore1, keyValueScore2)
 
     //case 4: key value scores with different keys but same values
     val keys1 = Array[Long](0, 1, 2)
     val keys2 = Array[Long](1, 2, 3)
     val sharedValues = Array[Double](1, 2, 3)
-    val keyValueScore3 = generateFullDataScores(keys1, sharedValues)
-    val keyValueScore4 = generateFullDataScores(keys2, sharedValues)
+    val keyValueScore3 = generateSimpleDataScores(keys1, sharedValues)
+    val keyValueScore4 = generateSimpleDataScores(keys2, sharedValues)
     assertNotEquals(keyValueScore3, keyValueScore4)
 
     //case 5: same keys and same values, but the order of key/value pairs are different
@@ -78,19 +78,19 @@ class ModelDataScoresTest extends SparkTestUtils {
     val keys4 = Array[Long](2, 1, 0)
     val values3 = Array[Double](4, 5, 6)
     val values4 = Array[Double](6, 5, 4)
-    val keyValueScore5 = generateFullDataScores(keys3, values3)
-    val keyValueScore6 = generateFullDataScores(keys4, values4)
+    val keyValueScore5 = generateSimpleDataScores(keys3, values3)
+    val keyValueScore6 = generateSimpleDataScores(keys4, values4)
     assertEquals(keyValueScore5, keyValueScore6)
   }
 
   @Test
   def testPlus(): Unit = sparkTest("testPlusForKeyValueScores") {
     //case 1: both key value scores are of length 0
-    val emptyScores = generateFullDataScores(scores = Array[Double]())
+    val emptyScores = generateSimpleDataScores(scores = Array[Double]())
     assertEquals(emptyScores - emptyScores, emptyScores)
 
     //case 2: one of the scores are empty
-    val nonEmptyScores = generateFullDataScores(scores = Array[Double](1, 2))
+    val nonEmptyScores = generateSimpleDataScores(scores = Array[Double](1, 2))
     assertEquals(emptyScores + nonEmptyScores, nonEmptyScores)
 
     //case 3: when both scores are non-empty
@@ -98,24 +98,24 @@ class ModelDataScoresTest extends SparkTestUtils {
     val keys2 = Array[Long](2, 3, 4)
     val values1 = Array[Double](1, -1, 1)
     val values2 = Array[Double](-1, 0, 1)
-    val keyValueScore1 = generateFullDataScores(keys1, values1)
-    val keyValueScore2 = generateFullDataScores(keys2, values2)
+    val keyValueScore1 = generateSimpleDataScores(keys1, values1)
+    val keyValueScore2 = generateSimpleDataScores(keys2, values2)
     val expectedKeys = Array[Long](0, 1, 2, 3, 4)
     val expectedValues = Array[Double](1, -1, 0, 0, 1)
-    val expectedKeyValueScore = generateFullDataScores(expectedKeys, expectedValues)
+    val expectedKeyValueScore = generateSimpleDataScores(expectedKeys, expectedValues)
     assertEquals(keyValueScore1 + keyValueScore2, expectedKeyValueScore)
   }
 
   @Test
   def testMinus(): Unit = sparkTest("testMinusForKeyValueScores") {
     //case 1: both key value scores are of length 0
-    val emptyScores = generateFullDataScores(scores = Array[Double]())
+    val emptyScores = generateSimpleDataScores(scores = Array[Double]())
     assertEquals(emptyScores - emptyScores, emptyScores)
 
     //case 2: one of the scores are empty
-    val nonEmptyScores = generateFullDataScores(scores = Array[Double](1, 2))
+    val nonEmptyScores = generateSimpleDataScores(scores = Array[Double](1, 2))
     assertEquals(nonEmptyScores - emptyScores, nonEmptyScores)
-    val expectedNonEmptyScores = generateFullDataScores(scores = Array[Double](-1, -2))
+    val expectedNonEmptyScores = generateSimpleDataScores(scores = Array[Double](-1, -2))
     assertEquals(emptyScores - nonEmptyScores, expectedNonEmptyScores)
 
     //case 3: when both scores are non-empty
@@ -123,15 +123,15 @@ class ModelDataScoresTest extends SparkTestUtils {
     val keys2 = Array[Long](2, 3, 4)
     val values1 = Array[Double](1, -1, 1)
     val values2 = Array[Double](-1, 0, 1)
-    val keyValueScore1 = generateFullDataScores(keys1, values1)
-    val keyValueScore2 = generateFullDataScores(keys2, values2)
+    val keyValueScore1 = generateSimpleDataScores(keys1, values1)
+    val keyValueScore2 = generateSimpleDataScores(keys2, values2)
     val expectedKeys1 = Array[Long](0, 1, 2, 3, 4)
     val expectedValues1 = Array[Double](1, -1, 2, 0, -1)
-    val expectedKeyValueScore1 = generateFullDataScores(expectedKeys1, expectedValues1)
+    val expectedKeyValueScore1 = generateSimpleDataScores(expectedKeys1, expectedValues1)
     assertEquals(keyValueScore1 - keyValueScore2, expectedKeyValueScore1)
     val expectedKeys2 = Array[Long](0, 1, 2, 3, 4)
     val expectedValues2 = Array[Double](-1, 1, -2, 0, 1)
-    val expectedKeyValueScore2 = generateFullDataScores(expectedKeys2, expectedValues2)
+    val expectedKeyValueScore2 = generateSimpleDataScores(expectedKeys2, expectedValues2)
     assertEquals(keyValueScore2 - keyValueScore1, expectedKeyValueScore2)
   }
 }
